Testing: Deprecated sql versions
